rename usethemebuilder file
