+++ apps/themebuilder/app/_components/appearance-toggle/appearance-toggle.tsx
@@ -4,7 +4,7 @@ import { MoonIcon, SunIcon } from '@navikt/aksel-icons';
 import cl from 'clsx/lite';
 import { useTranslation } from 'react-i18next';
 import { useSearchParams } from 'react-router';
-import { useThemebuilder } from '~/routes/themebuilder/_utils/useThemebuilder';
+import { useThemebuilder } from '~/routes/themebuilder/_utils/use-themebuilder';
 import classes from './appearance-toggle.module.css';
 
 type AppearanceToggleProps = {
