+++ apps/themebuilder/app/_components/color-contrasts/color-contrasts.tsx
@@ -13,7 +13,7 @@ import {
 import cl from 'clsx/lite';
 import { useState } from 'react';
 import { useTranslation } from 'react-i18next';
-import { useThemebuilder } from '~/routes/themebuilder/_utils/useThemebuilder';
+import { useThemebuilder } from '~/routes/themebuilder/_utils/use-themebuilder';
 import classes from './color-contrasts.module.css';
 
 const initialTheme = generateColorSchemes('#0062BA');
