+++ apps/themebuilder/app/_components/color-preview/color-preview.tsx
@@ -9,7 +9,7 @@ import {
 import cl from 'clsx/lite';
 import { type HTMLAttributes, useState } from 'react';
 import { useTranslation } from 'react-i18next';
-import { useThemebuilder } from '~/routes/themebuilder/_utils/useThemebuilder';
+import { useThemebuilder } from '~/routes/themebuilder/_utils/use-themebuilder';
 import listClasses from './card.module.css';
 import classes from './color-preview.module.css';
