+++ apps/themebuilder/app/_components/theme-header/theme-header.tsx
@@ -3,7 +3,7 @@ import { RRLink } from '@internal/components/src/link';
 import { ChevronLeftIcon } from '@navikt/aksel-icons';
 import { useTranslation } from 'react-i18next';
 import { useRouteLoaderData, useSearchParams } from 'react-router';
-import { useThemebuilder } from '~/routes/themebuilder/_utils/useThemebuilder';
+import { useThemebuilder } from '~/routes/themebuilder/_utils/use-themebuilder';
 import classes from './theme-header.module.css';
 
 export const ThemeHeader = () => {
